final version (semi-final)
if you run it, you find that there is nearly no change in values of pixels. 
I optimized the code by forcing the for loop to only visit indexes which are to be modified by including the interval in its advance. 
I only modified the blue value because it is unnecessary to modify the rest and i feel like blue is least noticeable in change. only the last 5 bits are modified, so the color remains virtually the same. there is an array for all keys accepted as inputs. 
I left some print statements in the loop so you can see what the color is.
To find out what the message is, you basically just go back through and get the lower 5 bits from the blue value.
This is basically done and is runnable