Update EncryptingText.pde
modified the code to hide the msg size within the first two pixels. this allows for a message size of over 1,000,000 characters if necessary (2^20).
the first pixel contains a marker character in the r value.
also modified the interval section of the code a bit. after further error checking, i realized there were several cases unaccounted for. modified the loop into a do while instead of a for loop because it seemed more appropriate and easier to read, given that multiple conditions needed to be checked and there were multiple loop variables.

added a section to check for markers at the beginning and end of an img. if the marker isnt found at the beginning, the body of the function is skipped and an error message is returned instead. if the marker isn't found at the end, an error message is returned then. otherwise it returns the message.